Fix the problem that some morphing data doesn't work properly

+++ src/foundation/materials/singles/MToonSingleMaterialNode.ts
@@ -326,9 +326,9 @@ export default class MToonSingleMaterialNode extends AbstractMaterialNode {
           initialValue: new Scalar(0), min: 0, max: Config.maxVertexMorphNumberInShader, needUniformInFastest: true
         },
         {
-          semantic: ShaderSemantics.DataTextureMorphOffsetPosition, componentType: ComponentType.Float, compositionType: CompositionType.ScalarArray, maxIndex: Config.maxVertexMorphNumberInShader,
+          semantic: ShaderSemantics.DataTextureMorphOffsetPosition, componentType: ComponentType.Int, compositionType: CompositionType.ScalarArray, maxIndex: Config.maxVertexMorphNumberInShader,
           stage: ShaderType.VertexShader, isSystem: true, soloDatum: true,
-          initialValue: new VectorN(new Float32Array(Config.maxVertexMorphNumberInShader)), min: -Number.MAX_VALUE, max: Number.MAX_VALUE, needUniformInFastest: true
+          initialValue: new VectorN(new Int32Array(Config.maxVertexMorphNumberInShader)), min: -Number.MAX_VALUE, max: Number.MAX_VALUE, needUniformInFastest: true
         },
         {
           semantic: ShaderSemantics.MorphWeights, componentType: ComponentType.Float, compositionType: CompositionType.ScalarArray, maxIndex: Config.maxVertexMorphNumberInShader,
